added indicator for r-b index.
