+++ xclim/indices/__init__.py
@@ -24,6 +24,7 @@ the implementation of the indices.
 """
 from ._anuclim import *
 from ._conversion import *
+from ._hydrology import *
 from ._multivariate import *
 from ._simple import *
 from ._threshold import *
